# backend/src/main/resources/api/api.yaml
openapi: 3.0.3
info:
  title: AIMS API
  description: API for AIMS backend
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/users/search:
    get:
      summary: Search users by login
      operationId: searchUsers
      tags:
        - users
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
        - name: role
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/RoleName"
      responses:
        200:
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSearchResponse"
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorObject"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorObject"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorObject"
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorObject"
  /api/v1/users/batch:
    post:
      summary: Batch load users by ids
      operationId: batchUsers
      tags:
        - users
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchUsersRequest"
      responses:
        200:
          description: Users found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchUsersResponse"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorObject"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorObject"
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorObject"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Login:
      type: string
      description: User login
      minLength: 1
      maxLength: 100
      example: "operator"
...
    RoleName:
      type: string
      description: Role name
      example: "OPERATOR"
...
    UserSummary:
      type: object
      required:
        - id
        - login
      properties:
        id:
          type: integer
          format: int64
        login:
          $ref: "#/components/schemas/Login"
    UserSearchResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/UserSummary"
    BatchUsersRequest:
      type: object
      required:
        - ids
      properties:
        ids:
          type: array
          maxItems: 100
          items:
            type: integer
            format: int64
    BatchUsersResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/UserSummary"
...
    ErrorObject:
      description: Error response object
      type: object
      properties:
        code:
          description: Numeric error code
          type: integer
          format: int32
        message:
          description: Unique text identifier of the error
          type: string
        humanMessage:
          description: Human-readable error description
          type: string